refactor: Rename files tftp_data_mgr* to tftpDataMgr* (u-t OK!)
Github issue #4
